requirements.txt
# =============================================
# Causal-MAC 项目依赖规范
# 版本: 1.0
# 更新日期: 2025-08-10
# =============================================

# ---------------------------------------------
# 核心框架
# ---------------------------------------------
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118
#torchaudio==2.1.0+cu118
# --extra-index-url https://download.pytorch.org/whl/cu118

# ---------------------------------------------
# 多智能体环境
# ---------------------------------------------
pettingzoo==1.24.0
gymnasium==0.29.1
pygame==2.5.0
supersuit==3.9.0

# ---------------------------------------------
# 数据处理与科学计算
# ---------------------------------------------
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.0
joblib==1.3.2
h5py==3.10.0

# ---------------------------------------------
# 因果推理
# ---------------------------------------------
causal-learn==0.1.4
dowhy==0.10
pgmpy==0.1.23

# ---------------------------------------------
# 模型优化与量化
# ---------------------------------------------
bitsandbytes==0.41.3
accelerate==0.24.1
transformers==4.35.2
peft==0.7.1

# ---------------------------------------------
# 配置管理与日志
# ---------------------------------------------
pyyaml==6.0.1
omegaconf==2.3.0
tqdm==4.65.0
tensorboard==2.14.1
wandb==0.16.0

# ---------------------------------------------
# 可视化
# ---------------------------------------------
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0
imageio==2.31.5

# ---------------------------------------------
# 开发与测试工具
# ---------------------------------------------
pytest==7.4.0
pytest-cov==4.1.0
coverage==7.3.0
mock==5.1.0
ipython==8.18.1
jupyter==1.0.0

# ---------------------------------------------
# 系统工具
# ---------------------------------------------
psutil==5.9.6
gpustat==1.1
nvidia-ml-py3==12.535.133

# ---------------------------------------------
# 其他依赖
# ---------------------------------------------
typing_extensions==4.8.0
numba==0.58.1
networkx==3.2.1
pydantic==2.5.3